pyproject.toml
[tool.poetry]
name = "m-remmitance"
version = "0.1.0"
description = "Mremmitance"
authors = ["ravi <ravi.sankoju@machint.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
starlette = "0.38.2"
aioredis = "2.0.1"
aiosmtplib = "3.0.2"
alembic = "1.13.2"
annotated-types = "0.7.0"
anyio = "4.4.0"
async-timeout = "4.0.3"
bcrypt = "3.1.7"
colorama = "0.4.6"
dnspython = "2.6.1"
email-validator = "2.2.0"
exceptiongroup = "1.2.2"
greenlet = "3.0.3"
h11 = "0.14.0"
httpcore = "1.0.5"
httptools = "0.6.1"
httpx = "0.27.0"
idna = "3.7"
importlib-metadata = "8.2.0"
importlib-resources = "6.4.0"
jinja2 = "3.1.4"
mako = "1.3.5"
markupsafe = "2.1.5"
mysql-connector-python = "9.0.0"
mysqlclient = "2.2.4"
passlib = "1.7.4"
phonenumbers = "8.13.43"
protobuf = "5.27.3"
pycparser = "2.22"
pydantic = "2.8.2"
pydantic-core = "2.20.1"
pyjwt = "2.9.0"
pymysql = "1.1.1"
python-dotenv = "1.0.1"
python-multipart = "0.0.9"
pyyaml = "6.0.2"
rfc3986 = "2.0.0"
six = "1.16.0"
sniffio = "1.3.1"
sqlalchemy = "2.0.32"
typing-extensions = "4.12.2"
uvicorn = "0.30.5"
zipp = "3.20.0"
jsonpickle = "^3.2.2"
fastapi = "^0.112.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
